pyproject.toml
[project]
name = "data-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "jinja2>=3.1.6",
    "polars>=1.33.1",
    "strawberry-graphql[cli]>=0.282.0",
    "uvicorn>=0.36.0",
]
